delete update of cc_pull,cc_push,pagerank_pull

+++ lonestar/analytics/distributed/sssp/sssp_pull.cpp
@@ -70,7 +70,7 @@ struct NodeData {
 
 galois::DynamicBitSet bitset_dist_current;
 
-typedef galois::graphs::DistGraph<NodeData, unsigned int> Graph;
+typedef galois::graphs::DistGraph<NodeData, double> Graph;
 typedef typename Graph::GraphNode GNode;
 
 std::unique_ptr<galois::graphs::GluonSubstrate<Graph>> syncSubstrate;
@@ -335,10 +335,10 @@ int main(int argc, char** argv) {
   std::unique_ptr<Graph> hg;
 #ifdef GALOIS_ENABLE_GPU
   std::tie(hg, syncSubstrate) =
-      distGraphInitialization<NodeData, unsigned int, false>(&cuda_ctx);
+      distGraphInitialization<NodeData, double, false>(&cuda_ctx);
 #else
   std::tie(hg, syncSubstrate) =
-      distGraphInitialization<NodeData, unsigned int, false>();
+      distGraphInitialization<NodeData, double, false>();
 #endif
 
   bitset_dist_current.resize(hg->size());
